requirements.txt
streamlit
llama-parse
llama-index
chromadb
google-generativeai
python-dotenv
nest-asyncio
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-vector-stores-chroma